#11 refactoring view recipe
