+++ src/components/ViewRecipe/ChipBar/DifficultyChip/DifficultyChip.test.js
