+++ src/components/ViewRecipe/ChipBar/RecipeChip/RecipeChip.js
@@ -1,6 +1,6 @@
 import { Chip, Tooltip } from "@mui/material";
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
-import { minutesToHours } from "../../../utils";
+import { minutesToHours } from "../../../../utils";
 
 export const RecipeChip = ({ value = "", label = "" }) => {
 	const valueToDisplay = value === "" ? "0:00" : minutesToHours(value);
