+++ src/components/ViewRecipe/RecipeHeader/RecipeImage/RecipeImage.test.js
