+++ src/components/ViewRecipe/index.js
@@ -1,2 +1,6 @@
 import { ViewRecipe } from "./ViewRecipe";
+import { ViewDetails } from "./ViewDetails/ViewDetails";
+import RecipeHeader from "./RecipeHeader";
+import { ChipBar } from "./ChipBar/ChipBar";
+export { ViewDetails, ChipBar, RecipeHeader };
 export default ViewRecipe;
